shift purge command to delete module
